+++ configs/harbor.toml
@@ -1,11 +1,11 @@
-# harbor-v1 — Terminal-Bench 2, each task run in its own declared container image (needs docker).
+# harbor — Terminal-Bench 2, each task run in its own declared container image (needs docker).
 # The `bash` harness gives the agent the shell tool it drives the terminal with.
 #
 #   uv run eval @ configs/harbor.toml
 num_tasks = 10
 
 [taskset]
-id = "harbor-v1"
+id = "harbor"
 dataset = "terminal-bench/terminal-bench-2"
 
 [harness]
