+++ environments/wordle_v1/wordle_v1/taskset.py
@@ -1,6 +1,6 @@
 """wordle-v1 — the textarena taskset pinned to Wordle (example env).
 
-A thin wrapper over `textarena_v1`: pins `game` to "Wordle-v0", so `wordle-v1` is a
+A thin wrapper over `textarena`: pins `game` to "Wordle-v0", so `wordle-v1` is a
 zero-config Wordle env. Everything else — the user simulator, seed-based task generation,
 and game-authoritative scoring — is inherited unchanged.
 """
@@ -8,7 +8,7 @@ and game-authoritative scoring — is inherited unchanged.
 from typing import Literal
 
 import verifiers.v1 as vf
-from verifiers.v1.tasksets.textarena_v1 import (
+from verifiers.v1.tasksets.textarena import (
     TextArenaConfig,
     TextArenaState,
     TextArenaTask,
