+++ verifiers/v1/tasksets/harbor/__init__.py
@@ -1,4 +1,4 @@
-from verifiers.v1.tasksets.harbor_v1.taskset import (
+from verifiers.v1.tasksets.harbor.taskset import (
     HarborConfig,
     HarborTask,
     HarborTaskset,
